Refactor help to not be a process module
The help process module always needed to have knowledge of the different
modules in the project, which was very awkward. This updates the help
system to be part of the process module and moves the help data into the
modules. This also allowed the removal of the visual boolean on the List
state, that was only used when rendering the help module.